+++ app/Jobs/StoreOneQuote.php
@@ -2,13 +2,13 @@
 
 namespace App\Jobs;
 
+use App\Classes\IexApi;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldBeUnique;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
-use App\Classes\StockQuote;
 
 class StoreOneQuote implements ShouldQueue
 {
@@ -36,7 +36,7 @@ class StoreOneQuote implements ShouldQueue
      */
     public function handle()
     {
-        $stock_quote_service = new StockQuote();
-        $stock_quote_service->store_one_quote($this->symbol, $this->date);
+        $iex_api = new IexApi();
+        $iex_api->store_one_quote($this->symbol, $this->date);
     }
 }