+++ app/Models/StockQuoteModel.php
@@ -27,7 +27,28 @@ class StockQuoteModel extends Model
         'pe_ratio',
         'week_52_low',
         'week_52_high',
+        'last_trade_time',
         'metadata',
+
+        // Testing / Development (today quote)
+//        'close_time',
+//        'delayed_price_time',
+//        'extended_price_time',
+//        'high_time',
+//        'iex_close_time',
+//        'iex_last_updated',
+//        'iex_open_time',
+//        'latest_time',
+//        'latest_update',
+//        'low_time',
+//        'open_time',
+
+
+        // Testing / Development (historic quote)
+//        'price_date',
+//        'date_date',
+//        'updated',
+//        'label',
     ];
 
     protected $casts = [